backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
sqlalchemy==2.0.25
google-generativeai==0.3.2
apscheduler==3.10.4
python-multipart==0.0.6
requests==2.31.0